Setup main.tf to call on module. Pending IAM module write.
Init basic files for a TF module

Base IAM role in the module to be called on later

Established basic module format. Have a placeholder for IAM module call in the main.tf

Added principals in tfvars so user can pick who they can assign IAM role to

Added developer and PM permissions in two separate data blocks to be used in the IAM module.

Added policy description to be populated into the IAM module